Refactored to be slightly more realistic
Before, everything went through the TodoDatabase type.  So much so, in fact, that
the actual todo item type, todoItem, didn't even need to be exported.  I felt this
was a bit unrealistic since other libraries are likely to give the todo item
a more 'first class' status.  So I exported TodoItem and made it part of the
TodoDatabase API.  Even so, the design is still very centered on the TodoDatabase
interface.  But I think this newer approach is probably more representative of
how 3rd party Go libraries might be coded.

This version has 98.1% code coverage.  I can't really get 100% because I cannot
force the uuid4 library to throw an error (which is the only thing that isn't
tested here).  I wish Golang had a pragma that allowed me to mark a line as
untestable.  Oh well.